user settings bug fixed.

+++ extensions/slideshare/templates/user-settings.php
@@ -42,26 +42,31 @@ $$username = bebop_tables::get_user_meta_value( $bp->loggedin_user->id, 'bebop_'
 $active = 'bebop_' . $extension['name'] . '_active_for_user';																//the active boolean name
 $$active = bebop_tables::get_user_meta_value( $bp->loggedin_user->id, 'bebop_' . $extension['name'] . '_active_for_user' );	//the value of the boolean
 
-echo '<form id="settings_form" action="' . $bp->loggedin_user->domain . 'bebop-oers/providers/?provider=' . $extension['name'] . '" method="post">
-<h3>' . ucfirst( $extension['name'] ) . ' Settings</h3>';
-if ( ! empty( $$username ) ) {
-	echo '<h5>Enable ' . ucfirst( $extension['name'] ) . ' import?</h5>
-	<input type="radio" name="bebop_' . $extension['name'] . '_active_for_user" id="bebop_' . $extension['name'] . '_active_for_user" value="1"';  if ( $$active == 1 ) {
-		echo 'checked';
-	} echo '>
-	<label for="yes">Yes</label>
-	<input type="radio" name="bebop_' . $extension['name'] . '_active_for_user" id="bebop_' . $extension['name'] . '_active_for_user" value="0"'; if ( $$active == 0 ) {
-		echo 'checked';
-	} echo '>
-	<label for="no">No</label><br>';
+if ( ( bebop_tables::get_option_value( 'bebop_' . $extension['name'] . '_provider' ) == 'on') && ( bebop_tables::check_option_exists( 'bebop_' . $extension['name'] . '_consumer_key' ) ) ) {
+	echo '<form id="settings_form" action="' . $bp->loggedin_user->domain . 'bebop-oers/providers/?provider=' . $extension['name'] . '" method="post">
+	<h3>' . ucfirst( $extension['name'] ) . ' Settings</h3>';
+	if ( ! empty( $$username ) ) {
+		echo '<h5>Enable ' . ucfirst( $extension['name'] ) . ' import?</h5>
+		<input type="radio" name="bebop_' . $extension['name'] . '_active_for_user" id="bebop_' . $extension['name'] . '_active_for_user" value="1"';  if ( $$active == 1 ) {
+			echo 'checked';
+		} echo '>
+		<label for="yes">Yes</label>
+		<input type="radio" name="bebop_' . $extension['name'] . '_active_for_user" id="bebop_' . $extension['name'] . '_active_for_user" value="0"'; if ( $$active == 0 ) {
+			echo 'checked';
+		} echo '>
+		<label for="no">No</label><br>';
+	}
+	
+	echo '<label for="bebop_' . $extension['name'] . '_username">' . ucfirst( $extension['name'] ) . ' Username:</label>
+	<input type="text" name="bebop_' . $extension['name'] . '_username" value="' . $$username .'" size="50"><br>
+	
+	<div class="button_container"><input type="submit" class="standard_button" value="Save Settings"></div>';
+	if ( ! empty( $$username ) ) {
+		echo '<div class="button_container"><a class="standard_button" href="?provider=' . $extension['name'] . '&reset=true">Remove Channel</a></div>';
+	}
+	echo '</form>';
 }
-
-echo '<label for="bebop_' . $extension['name'] . '_username">' . ucfirst( $extension['name'] ) . ' Username:</label>
-<input type="text" name="bebop_' . $extension['name'] . '_username" value="' . $$username .'" size="50"><br>
-
-<div class="button_container"><input type="submit" class="standard_button" value="Save Settings"></div>';
-if ( ! empty( $$username ) ) {
-	echo '<div class="button_container"><a class="standard_button" href="?provider=' . $extension['name'] . '&reset=true">Remove Channel</a></div>';
+else {
+	echo ucfirst( $extension['name'] ) . ' has not yet been configured. Please contact the blog admin to make sure ' . ucfirst( $extension['name'] ) . ' is configured properly.';
 }
-echo '</form>'; 
 ?>
